Let the remote tracking branch be configurable
This commit allows a user to define a per repository or per branch
specific configuration of which branch git-radar compares it to. If no
branch is configured explicitly "origin/master" will be used (as was
until this feature).

A branch specific remote tracking branch may be configured by executing

    git config --local branch."$(git rev-parse --abbrev-ref HEAD)".git-radar-tracked-remote origin/my-branch-to-compare-to

A repository specific remote tracking may be configured by executing

    git config --local git-radar.tracked-remote origin/my-branch-to-compare-to

A branch specific will overwrite a repository specific configuration
which will - naturally - overwrite the default.